move jpsurv functions
